# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agentic Legacy Modernization System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface IngestRequest {
  repoUrl: string;
}

export interface IngestResponse {
  jobId: string;
  projectId: string;
  projectName: string;
  fileCount: number;
}

export interface Project {
  id: string;
  repo_url: string;
  name: string;
  status: string;
  file_count: number;
  created_at: string;
  updated_at: string;
}

export interface AnalyzeRequest {
  projectId: string;
}

export interface ApiRoute {
  id: string;
  project_id: string;
  method: string;
  path: string;
  description?: string | null;
  handler?: string | null;
  created_at: string;
}

export interface AnalyzeResponse {
  jobId: string;
  projectId: string;
  apiCount: number;
  apis: ApiRoute[];
}

export interface GeneratePRDRequest {
  projectId: string;
}

export interface GeneratePRDResponse {
  jobId: string;
  documentId: string;
  projectId: string;
  title: string;
}

export interface Document {
  id: string;
  project_id: string;
  type: string;
  title: string;
  content: string;
  created_at: string;
}

export interface ExtractDbSchemaRequest {
  projectId: string;
}

export interface DbColumn {
  id: string;
  name: string;
  type: string;
  is_primary: boolean;
  is_nullable: boolean;
}

export interface DbTable {
  id: string;
  name: string;
  extracted_at: string;
  columns: DbColumn[];
}

export interface DbFunction {
  id: string;
  name: string;
  parameters?: string | null;
  description?: string | null;
}

export interface ExtractDbSchemaResponse {
  jobId: string;
  projectId: string;
  tableCount: number;
  functionCount: number;
  tables: DbTable[];
  functions: DbFunction[];
  extractedAt: string;
}

export interface DbSchemaResult {
  projectId?: string;
  tables: DbTable[];
  functions: DbFunction[];
  extractedAt?: string | null;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface Job {
  id: string;
  agentType: string;
  projectId?: string | null;
  status: JobStatus;
  message: string;
  startedAt: string;
  completedAt?: string | null;
  error?: string | null;
}

export interface AstTestRequest {
  /** File path including extension (e.g. src/foo.java) */
  filePath: string;
  /** Source code to parse */
  code: string;
}

export interface AstChunkMetadata {
  type: string;
  name: string;
  file: string;
  language: string;
  route?: string | null;
  method?: string | null;
  lineStart?: number | null;
  lineEnd?: number | null;
  className?: string | null;
}

export interface AstChunkItem {
  id: string;
  content: string;
  metadata: AstChunkMetadata;
}

export interface AstTestResponse {
  filePath: string;
  language: string;
  chunkCount: number;
  chunks: AstChunkItem[];
}

export type SearchResultItemMetadata = {
  type: string;
  file?: string | null;
};

export interface SearchResultItem {
  id: string;
  content: string;
  metadata: SearchResultItemMetadata;
  score: number;
}

export interface SearchResponse {
  projectId: string;
  query: string;
  indexedDocuments: number;
  results: SearchResultItem[];
}

export type ListProjects200 = {
  projects: Project[];
};

export type ListApis200 = {
  apis: ApiRoute[];
};

export type ListDocuments200 = {
  documents: Document[];
};

export type ListJobs200 = {
  jobs: Job[];
};

export type SemanticSearchParams = {
  /**
   * Project ID to search within
   */
  projectId: string;
  /**
   * Natural language search query
   */
  q: string;
  /**
   * Max results to return (1–20)
   */
  n?: number;
};