# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agentic Legacy Modernization System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/ingest:
    post:
      operationId: ingestRepository
      tags: [ingestion]
      summary: Ingest a repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequest"
      responses:
        "201":
          description: Ingestion started and completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /agent/projects/{id}:
    get:
      operationId: getProject
      tags: [ingestion]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    IngestRequest:
      type: object
      properties:
        repoUrl:
          type: string
          format: uri
      required: [repoUrl]
    IngestResponse:
      type: object
      properties:
        jobId:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        fileCount:
          type: integer
      required: [jobId, projectId, projectName, fileCount]
    Project:
      type: object
      properties:
        id:
          type: string
        repo_url:
          type: string
        name:
          type: string
        status:
          type: string
        file_count:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, repo_url, name, status, file_count, created_at, updated_at]
...